+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed an issue which caused glTF loading function to return before all resources was loaded.
+
 ## [1.6.0] - 2022-08-14
 ### Added
 - Added support for PixiJS extensions API.